FIX generateGuiaDespachoPDF
Ahora la funcion generateGuiaDespachoPDF esta en su propio archivo y es consultada desde pdf.ts
npm run build exitoso y todo funcionando correctamente
